First version of https://www.edb.cz/katalog-firem/ sampling
Implement scraping and sampling of the https://www.edb.cz/katalog-firem/
catalog. There are three versions of the "kontakt" page, which we have
to distinguish and scrape each one differently. There might be more but
the 3 are all we have seen so far. We scrape the company name, company
email address or addresses, the industry the company is from and the url
to the contacts page.

If company has multiple emails, it gets multiple rows in the csv.