Added group cache, fixed retry, added group info
Added group cache, so that the number of pages for each group does not
have to be determined when restarting the scraping script, as it
increases the number of requests. The group cache is dumped after each
run or when the run is interrupted.

Fixed retry, so that for small groups with many companies without
emails, we just retry sampling some other group rather than fishing for
a company with an email in the given group.

Added pyproject.toml with required dependencies