Allow multiple files when loading exclude set, make cache validation optional
Allow multiple files when loading exclude set, so that we can exclude
both previous results and current already collected results.

Make cachce validation optional, as it takes a long time and most groups
do not change in a short span of time.

Optimize num groups listing, so that it does not do cache validation and
stuff, minimizing number of requests.

Add command to fill the group cache separately from scraping, so that
these two actions can be done independently.